ARM64: mmu: implement mmu_disable completely in assembly
Splitting mmu_disable into two noinline function on a RK3568 leads to a
barebox crash, because the code has the implicit assumption that the
compiler won't generate memory accesses including spilling to stack.

We can't guarantee this in C code, so implement the procedure in
assembly.

While at it, drop mmu_early_disable(), which is unused, superfluous and
suffers from the same issue as old mmu_disable().

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
---
v2 -> v3:
  - use switch_el (Sascha)
  - fix typo: s/prologue/epilogue/ at the end
v1 -> v2:
  - fix copy-paste problem in some ELs (Ulrich)