efi: loader: move PE implementation out of common code
The PE code was initially added to a common location to mimic the ELF
code (which is used across architectures and also for remoteproc).

Since then ELF handling has been moved out of common/bootm.c and I
reached the conclusion that there is little benefit of supporting
running PE executables without UEFI.

Therefore move the implementation to efi/loader/ instead and give it an
API suitable for use from the EFI loader.

The code was not in use before, so we just use the occasion to sync with
U-Boot.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>