efi: loader: add object handling API
Add definitions and prototypes for the EFI object model.
The implementations will follow later, but we add this first here, so we
do not have to add everything in one big commit and can split it out
into separate parts.

The code can still be compiled, but it will only be linkable with linker
garbage collection enabled.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>