efi: loader: CONFIG_EFI_RT_VOLATILE_STORE
If we store EFI variables on ESP during boottime, we need cooperation
from the OS to write that file to implement SetVariable.

Populate the necessary options to allow efivar(1) to read a dump
of the variables after writing them and then persist them to ESP.

Note that this violates the EFI spec and it's not power-fail safe, but
it's an ok fallback for now.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>